refactor: extract draft actions into separate hook for improved modularity
This commit introduces a new useDraftActions hook that centralizes draft event management actions, further modularizing the draft event handling logic. Key changes include:

- Extracting complex draft-related methods from useDraft
- Improving separation of concerns
- Simplifying the main draft hook's implementation
- Maintaining consistent state management across draft interactions
