+++ packages/web/src/views/Calendar/components/Grid/AllDayRow/AllDayRow.tsx
@@ -20,6 +20,8 @@ import { isEventFormOpen } from "@web/common/utils";
 import { StyledGridCol } from "../Columns/styled";
 import { StyledAllDayColumns, StyledAllDayRow } from "./styled";
 import { AllDayEvents } from "./AllDayEvents";
+import { Schema_GridEvent } from "@web/common/types/web.event.types";
+import { selectIsDrafting } from "@web/ducks/events/selectors/draft.selectors";
 
 interface Props {
   dateCalcs: DateCalcs;
@@ -40,13 +42,32 @@ export const AllDayRow: FC<Props> = ({
 
   const { startOfView, weekDays } = weekProps.component;
   const rowsCount = useAppSelector(selectRowCount);
+  const isDrafting = useAppSelector(selectIsDrafting);
 
   useEffect(() => {
     measurements.remeasure(ID_GRID_MAIN);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [rowsCount]);
 
-  const startAlldayDraft = async (e: MouseEvent) => {
+  const openAlldayDraft = async (e: MouseEvent, event: Schema_GridEvent) => {
+    e.stopPropagation();
+
+    if (isDrafting) {
+      console.log("todo: close draft");
+    } else {
+      console.log("opening draft for:", event.title);
+      dispatch(
+        draftSlice.actions.start({
+          activity: "gridClick",
+          eventType: Categories_Event.ALLDAY,
+          event,
+        })
+      );
+    }
+  };
+
+  const startNewAlldayDraft = async (e: MouseEvent) => {
+    //TODO move this to mouse handler?
     const x = getX(e, isSidebarOpen);
     const startDate = dateCalcs.getDateStrByXY(
       x,
@@ -73,14 +94,14 @@ export const AllDayRow: FC<Props> = ({
       return;
     }
 
-    await startAlldayDraft(e);
+    await startNewAlldayDraft(e);
   };
 
   return (
     <StyledAllDayRow
       id={ID_GRID_ALLDAY_CONTAINER}
       rowsCount={rowsCount}
-      onMouseDown={onSectionMouseDown}
+      // onMouseDown={onSectionMouseDown}
     >
       <StyledAllDayColumns id={ID_ALLDAY_COLUMNS} ref={allDayRef}>
         {weekDays.map((day) => (
@@ -91,6 +112,7 @@ export const AllDayRow: FC<Props> = ({
         measurements={measurements}
         startOfView={weekProps.component.startOfView}
         endOfView={weekProps.component.endOfView}
+        onClick={openAlldayDraft}
       />
     </StyledAllDayRow>
   );
